Move Gui-objects to seperate file Add datafile load option
